Added support for specifying an arbitrary GBNF compatible grammar in the Modelfile, for models running on the llama.cpp backend
Note that this is basically just the same PR as the one submitted
by SyrupThinker in September (#565), and that has been mentioned
in issue #1507 and #808 since then.

There are plenty of users that would appreciate this feature, so
I really hope that it can get merged.

It's great that support for JSON grammar specifically has been
added, by setting the GBNF grammar in question when JSON format
is requested, but by providing the user with the ability to
specify an arbitrary grammar opens up for a lot more possibilities
than that

Pull request #830 adds support for specifying JSON schemas, which
is yet another great convenience feature for a specific and common
usecase, but by adding support for arbitrary GBNF grammar it would
be possible to have any model outputting data in any type of format,
including custom DSLs and text-based file formats in general

This is a tremendously useful thing to have when building various
types of automation related applications, so I really hope that this
can get merged to avoid having to maintain separate forks. Ollama
is a great project, let's keep making it even better