fix pasting slash commands
there is a bug in paste where the pasted content is written directly to
the prompt buffer instead of being processed. for most content, this is
fine but slash commands are processed line-by-line.

aggregate status updates, e.g. "Set 'verbose' mode.", "Set system
message.", to the end for aesthetics. the status message shouldn't
display while in paste mode