Introduce GBNF grammar parameter: run flags, REPL and API
This patch adds full support for the llama.cpp "grammar" option, which
defaults to "" i.e. no grammar constraint.

For backwards compatability, the existing `format` parameter remains
available in flags, the REPL and the API. Only one of `grammar` or
`format` at a time is legal, since `format: json` is just a shortcut
for the hard-coded JSON grammar.

Tested locally.