Install Ollama on OSTree systems
There's a large plethora of OSTree OSes in the Fedora family:

Silverblue, Kinoite, CoreOS, IoT, Onyx, Sericea, Vauxite

In the CentOS Stream family:

Automotive Stream Distribution, CoreOS

In the Red Hat family:

Red Hat In-Vehicle Operating System, Red Hat Enterprise Linux CoreOS,
RHEL for Edge

Then there's the Universal Blue family:

Things like podman-machine and podman-desktop on Windows and macOS use
Fedora CoreOS as the host OS so there is that also.

The list goes on and on.

These OSes are ideal for containerized AI LLMs like Ollama.

I eventually got this working with podman, which is probably the best
route to use on these OSes (or an rpm when someone packages it).

This change gets the install script working.

/usr/bin and /usr/share aren't writable on these systems, but
/usr/local/bin and /usr/local/share are.

So this change ensures if /usr/local/bin is used during installation,
that we also use /usr/local/share, if /usr/local/share exists. And then
everything seems to work fine on these OSes for non-containerized
installs.

Signed-off-by: Eric Curtin <ecurtin@redhat.com>