server/routers.go: Fix checkNameExists
When copy a model with a existed name, it suppose to error out,
but it success. The old `checkNameExists` only report exist model
name in case the name is different, but they are same when both of
them convert to upper case.

```
ollama cp tinyllama tinyllama
copied 'tinyllama' to 'tinyllama'
```

And we should not check name existence on create and pull, because
it suppose to replace some layers of the existing model

Signed-off-by: Lei Jitang <leijitang@outlook.com>