Check ROCM install for libhipblas.so* instead of libhipblas.so.2*
AMDGetGPUInfo previously required libhipblas.so.2* to be present
in the ROCM installation or it would ignore availalbe GPUs.
libhipblas 2 is only available starting from ROCm v6.
Ollama appears to be compiling and running fine with older versions
of ROCm (tested for 5.6.1) which come with libhipblas.so.1*.

This commit relaxes the filter in ROCmLibGlobs to accept an
libhipblas.so*.