Index only available AMD GPUs.
Previously AMDGetGPUInfo assigns GPU IDs for each node in
/sys/class/kfd/kfd/topology/nodes/ and expects these
to map directly to IDs as used in HIP/ROCR_VISIBLE_DEVICES.

This is not always the case, e.g., when used with the SLURM
scheduler, which will hide devices from jobs using cgroups.
In that case IDs assumed by Ollama and assigned by ROCr
will differ, causing Ollama to fail to start worker processes
(or applying wrong filters based on HIP/ROCR_VISIBLE_DEVICES).

We can detect which nodes are hidden from the current process
since trying to read the corresponding files in
/sys/class/kfd/kfd/topology/nodes/... results in permission
errors.

This commit adjusts the GPU ID assignment in AMDGetGPUInfo
to only assign an ID to devices/nodes that are accessible
(not skipped due to read permission errors of the properties files)
in an incremental fashion.

Caveat: This works on the tested system, but there may be
edge cases in which it fails.