feat: Add config plumbing for TLS server options
(m)TLS config is a set of three values for both client-side and server-side
but they're asymmetrical. For standard TLS, the server needs a key/cert and
the client needs the CA IFF the server's key/cert is signed by a non-system
CA. For mTLS, the server needs to add the CA used to sign the client's key/
cert pair, and the client needs a key/cert pair.

Currently, it is assumed that for mTLS, the CA to verify the client's cert
will be signed by a non-system CA since this is typically the case.

https://github.com/ollama/ollama/issues/3203

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>