Enable server to be activated via systemd sockets.
Systemd sockets start the corresponding service when there is a request
on the specified port. Systemd will then forward the port to the service
by passing a file descriptor. The service must be prepared to accept
this, since it cannot just bind to the same port the socket is already
bound to. This change enables this.

That this works can be tested via
```
systemd-socket-activation -l 11434 ./ollama server
```

This will bind to port 11434 but not start ollama. When trying to
connect to this port, the server will actually start.