runner.go: Improve health status reporting when laoding model
Currently, models are loaded synchonously before the HTTP server
is created. This means that the health endpoint is not really
useful - the only status it can report is fully loaded. For
large models on slow storage, this can result in timeouts since
Ollama doesn't see any progress. This moves model loading into
a separate goroutine so that the HTTP server can respond to
requests while loading is happening.

In addition, it is not safe to write and access strings for status
across threads - this switches to ints, which are atomic.