llama.go: Pass through additional context params
Ollama with the current server.cpp backend initializes batch size
and max seqs when loading models, whereas the current Go backend
leaves them as the defaults.

The only direct problem this causes is if batch size is set to
larger than the default through Ollama. Otherwise the default
values mostly just waste memory or leave a surprise for the
future that will be hard to track down.