Make libhipblas version runtime check compile-time configurable.
AMDGetGPUInfo requires libhipblas.so.2* to be present by default
in the ROCM installation or it would ignore availalbe GPUs.
libhipblas 2 is only available starting from ROCm v6.
Ollama appears to be compiling and running fine with older versions
of ROCm (tested for 5.6.1) which come with libhipblas.so.1*.

However, relaxing the filter in ROCmLibGlobs to accept either version
could lead into problems where Ollama is compiled against one version
of ROCm but finds another that is incompatible during runtime.

As a soluation, the library name to search for is now stored in
variable gpu.ROCmLib and can be adjusted during compilation using
export GOFLAGS="'-ldflags=-w -s
\"-X=github.com/ollama/ollama/gpu.ROCmLib=libhipblas.so.1*\"'".

This approach was suggested by Daniel Hiltgen (dhiltgen).