runner.go: Don't panic when processing sequences
If there is an error processing a sequence, we should simply return
an HTTP error and abort that sequence rather than panic the whole
runner. This will make us more resilient to transient failures.

Panics can still occur during startup as there is no way to serve
requests if that fails.

Based on some code that was originally part of the vision work.

Co-authored-by: jmorganca <jmorganca@gmail.com>