Add support for v5 and v6+ multi-gpu cuda setups
With the new addition of cuda v12 and selecting the newer version for GPUs
that support it, users with a mix of v11 only and v12 supporting GPUs weren't
able to load models spanning the GPU.  This refines the algorithm to use the oldest
variant (v11 in this case) so that all the GPUs will be supported.  This will break
down if a user has a CC 5.x card and a CC 9.x card, as we don't currently
have plumbing in place to understand the maximum version support.  The Go
server will open up the ability to expose a more refined support matrix from
the runners, and we can refine this algorithm then.