cuda: adjust variant based on detected runners
When building from source, or using downstream packaging systems, multiple
versions of cuda runners may not be present.  This adjusts the discoverry logic
to only use versioned variants if they are detected at runtime.  It also adds a new
warning message in the log if no cuda runners are present but cuda GPUs are
detected.