ggml-backend.go: Don't do async computation when returning data
There are two potential issues with async computation when we need
the resulting data:
 - The computation may not have completed by the time we need the data
 - The Go buffer could move while the computation is happening

When we are doing this type of operation, getting the data is the
very next step after computation so there really isn't a benefit
to having an async call, so we might as well just do a blocking
computation and avoid the above problems.

If there is no data to be returned then the problems don't apply
and so we should not block on the operation completing.