server/internal/client/ollama: persist through chunk download errors
This changes the behavior of the registry client to persist through
chunk download errors. The client will no longer bring down all
in-flight chunk downloads if one of them fails. Instead, it will
continue to download the other chunks, and once all of them are
complete, it will return an error if any of them failed. Clients can use
the tracing feature to get updates on the progress of each chunk and
cancel their context if they want to stop the download early.

This also changes the progress reporting behavior. Now, the progress
reporting will report, per read, the number of bytes received for the
read. Previously, it would report the total number of bytes received for
the entire blob. This caused issues when the blob was downloaded in a
single chunk, as the progress would not be updated until the entire blob
was downloaded.

Also, provide an example of how to use the tracing feature to cancel the
download early should clients want to do that.

Also, using 'go run .' now sets the Ollama version "v0.0.0" instead of
"(devel)" to avoid issues with the registry server that were encountered
and blocking some testing of this change.