server/internal/client/ollama: handle some network errors gracefully
In some cases, the client may encounter network errors that are
temporary in nature, such as connection resets or timeouts. In these
cases, the client should retry the operation instead of failing
immediately.

Hopefully this helps users who are experiencing intermittent network
errors in the middle of large downloads, causing it to fail, possible
unnoticed by the user for some time; time that could be spent resuming
the download on their behalf.

This also updates the DefaultRegistry docs to mention the new
ReadTimeout field, which it sets to 30 seconds.

Also, remove old mention of removed ChunkingDirectory field.