Make download resuming possible on large files
This change modifies the pattern to cURL and pipe into TAR, so that cURL
resumption on network error works properly. The new pattern is for cURL to
download to an actual (resumable) file.

Three spots affected in the install script, assumedly the only three large
file downloads.

The cleanup logic is also extended to keep files around in a temp directory,
until full success of the script. The temp directory needs a predictable
name, so the script does not use `mktemp` anymore.