fix: Use F32 for recurrent layer size estimates
In llama.cpp, all recurrent and hybrid-recurrent caching uses F32. It is
technically configurable in the memory, but there are no places where it is
set to anything besides F32. If this changes, it will look like a special
case in llama_model::create_memory, so we'll need a corresponding special
case here.

Branch: GraniteHybridMemoryEstimates

Signed-off-by: Gabe Goodhart <ghart@us.ibm.com>