Support overriding tensor-split from config
When it comes to automatic determination of how to optimally split a model
that doesn't fit into VRAM of a single GPU, ollama does a rather poor job.
Up until 0.11.4, the binary could be wrapped with a shell script and the
layer split parameters overridden to manually optimize the layer split across
GPUs, but this went away with 0.11.5.

How bad is the problem? For example, for llama 3.x 70b models and variants thereof
(e.g. hermes 4), ollama offloads a total of 19/81 layers to my 4x22GB GPUs with
context set to maximum 128K. With the manual override configured to 18,21,21,21 81/81
layers are offloaded, still with the context maxed out at 128K, without OOMs, tested up
to full 128K of context.

While evolving better default heuristics is appreciated, it seems very unrealistic to
expect that they will ever be perfect for every model, and when they are not, there is
currently no way to override the split to optimize it. hermes4:70b 19/81 with layers
offloaded is completely unusable, but with 81/81 in VRAM across the 4 GPUs, it is very quick.
This patch introduces a clean feature to optionally explicitly override the layer split on a
model-by-model basis using overrides stored in an .ini file (defaults to ~/.ollama.ini)
If the file is present (config file path/name can be specified using OLLAMA_OVERRIDE_CONFIG
if anyone deems the default unsuitable for any reason), it will be used. If it is not, or the
requested model doesn't have an override config specified in the .ini file, default heuristic
will be used. Example config block in the .ini file:

```
[hermes4:70b]
tensor-split=18,21,21,21
```

Test cases are also included for regression testing against future updates.